Add Binary Strings
Problem Description

Given two binary strings A and B. Return their sum (also a binary string).

Problem Constraints

1 <= length of A <= 105

1 <= length of B <= 105

A and B are binary strings

Input Format

The two argument A and B are binary strings.

Output Format

Return a binary string denoting the sum of A and B

Example Input

Input 1:
A = "100"
B = "11"
Input 2:
A = "110"
B = "10"

Example Output

Output 1:
"111"
Output 2:
"1000"

Example Explanation

For Input 1:
The sum of 100 and 11 is 111.
For Input 2:

The sum of 110 and 10 is 1000.

Signed-off-by: Surya V <suryabluz007@gmail.com>